Fixed windows circular import
Moved all modifications to modules not needed in arch that were causing
crashed to another file next to __init__ (in arch.windows).
